+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/base/DatabaseVariantContextProvider.java
@@ -9,6 +9,7 @@ import htsjdk.variant.variantcontext.VariantContext;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 public interface DatabaseVariantContextProvider {
 
 	/**