+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/base/VCFReaderVariantProvider.java
@@ -11,6 +11,7 @@ import java.io.File;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 public class VCFReaderVariantProvider implements DatabaseVariantContextProvider {
 
 	/**