+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/base/VariantNormalizer.java
@@ -16,6 +16,7 @@ import java.io.FileNotFoundException;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 public final class VariantNormalizer {
 
 	/**