+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/clinvar/ClinVarAnnotation.java
@@ -11,6 +11,7 @@ import java.util.Collection;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 public class ClinVarAnnotation {
 
 	/**