+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/dbsnp/DBSNPRecordBuilder.java
@@ -8,6 +8,7 @@ import java.util.HashSet;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 public class DBSNPRecordBuilder {
 
 	/**