+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/generic_tsv/GenericTSVHeaderExtender.java
@@ -10,6 +10,7 @@ import htsjdk.variant.vcf.VCFInfoHeaderLine;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 public class GenericTSVHeaderExtender extends VCFHeaderExtender {
 
 	private GenericTSVAnnotationOptions tsvOptions;