+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/generic_vcf/GenericVCFVariantContextToRecordConverter.java
@@ -8,6 +8,7 @@ import htsjdk.variant.variantcontext.VariantContext;
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
  */
+@Deprecated
 final class GenericVCFVariantContextToRecordConverter implements VariantContextToRecordConverter<VariantContext> {
 
 	@Override