+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/gnomad/GnomadRecordBuilder.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
+@Deprecated
 public class GnomadRecordBuilder {
 
 	/**